Better warning fix

+++ Items.h
@@ -191,15 +191,15 @@ public:
 	DEPRECATE("Use GetSlot(0) instead of Slot1")
 	inline int getSlot1() const { return m_slots[0]; }
 	DEPRECATE("Use SetSlot(0) instead of Slot1")
-	inline void setSlot1(int v) { m_slots[0] = v; }
+	inline void setSlot1(int v) { m_slots[0] = static_cast<short>(v); }
 	DEPRECATE("Use GetSlot(1) instead of Slot2")
 	inline int getSlot2() const { return m_slots[1]; }
 	DEPRECATE("Use SetSlot(1) instead of Slot2")
-	inline void setSlot2(int v) { m_slots[1] = v; }
+	inline void setSlot2(int v) { m_slots[1] = static_cast<short>(v); }
 	DEPRECATE("Use GetSlot(2) instead of Slot3")
 	inline int getSlot3() const { return m_slots[2]; }
 	DEPRECATE("Use SetSlot(2) instead of Slot3")
-	inline void setSlot3(int v) { m_slots[2] = v; }
+	inline void setSlot3(int v) { m_slots[2] = static_cast<short>(v); }
 };
 
 
@@ -236,7 +236,7 @@ public:
 
 	inline ItemGlobalIndex(ItemContainerInstance location, int slot1, int slot2 = -1, int slot3 = -1)
 		: Location(location)
-		, Index(slot1, slot2, slot3)
+		, Index(static_cast<short>(slot1), static_cast<short>(slot2), static_cast<short>(slot3))
 	{}
 
 	inline void SetLocation(ItemContainerInstance location) { Location = location; }
